Add form components: HiddenField, DateField, StringField, and FormFieldTitle; refactor routes to use new TaxonomicForm component
