+++ src/components/general/FormComponents/TaxonomicForm.tsx
@@ -3,22 +3,34 @@ import { useState } from 'react';
 import { Container, Row, Col, Card } from 'react-bootstrap';
 
 /* Import Sources */
+import TaxonomicExpertFormJSON from 'sources/forms/TaxonomicExpertForm.json';
 import TaxonomicServiceFormJSON from 'sources/forms/TaxonomicServiceForm.json';
 
 /* Import Components */
 import Header from 'components/general/header/Header';
 import Footer from 'components/general/footer/Footer';
-import FormBuilder from './taxonomicServiceFormComponents/FormBuilder';
 import { BreadCrumbs } from 'components/general/CustomComponents';
+import FormBuilder from 'components/general/FormComponents/FormBuilder';
+import { Color, getColor } from '../ColorPage';
+
 
 
 /**
  * Component that renders the taxonomic service form
  * @returns JSX Component
  */
-const TaxonomicServiceForm = () => {
+const TaxonomicForm = () => {
     /* Base variables */
     const [completed, setCompleted] = useState<boolean>(false);
+
+    const title = location.pathname.includes("/te") ? 'Register as a new taxonomic expert' : 'Suggest a new taxonomic e-service';
+    const description = location.pathname.includes("/te") ? 'Use this form to register as a new expert that should be listed in the CETAF Marketplace.\nPlease fill in the required fields and add as much additional information as you can.\nThe CETAF secretariat will review and curate your submission before adding it to the marketplace\nTo be accepted it needs to be a taxonomic tool or service and it needs to be of sufficient quality.\nPlease propose only services that are in production and maintained.' : 'Use this form to suggest new taxonomic e-services or tools that should be listed in the CETAF Marketplace.\nPlease fill in the required fields and add as much additional information as you can.\nThe CETAF secretariat will review and curate your submission before adding it to the marketplace.\nTo be accepted it needs to be a taxonomic tool or service and it needs to be of sufficient quality\nPlease propose only services that are in production and maintained.';
+    const sended = 'Your submission is received and will be processed by the Marketplace! A CETAF administrator will review and score the taxonomic service. When the score is sufficient, the service will be published in the Marketplace catalog.\nThank you for using the Taxonomic Marketplace!';
+    
+    const formTemplate = location.pathname.includes("/te") ? TaxonomicExpertFormJSON : TaxonomicServiceFormJSON;
+
+    /* Determine color */
+    const color = "fs-2 tc-" + getColor(window.location) as Color;
 
     return (
         <div className="h-100 d-flex flex-column">
@@ -43,26 +55,18 @@ const TaxonomicServiceForm = () => {
                                     {/* Form title and description */}
                                     <Row>
                                         <Col>
-                                            <h1 className="fs-2 tc-primary">
+                                            <h1 className={color}>
                                                 {!completed ?
-                                                    'Suggest a new taxonomic e-service'
+                                                    title
                                                     : 'Submission received'
                                                 }
                                             </h1>
                                             {!completed ?
                                                 <p className="mt-3 fs-4">
-                                                    Use this form to suggest new taxonomic e-services or tools that should be listed in the CETAF Marketplace.
-                                                    Please fill in the required fields and add as much additional information as you can.
-                                                    The CETAF secretariat will review and curate your submission before adding it to the marketplace.
-                                                    To be accepted it needs to be a taxonomic tool or service and it needs to be of sufficient quality.
-                                                    Please propose only services that are in production and maintained.
+                                                    {description}
                                                 </p>
                                                 : <p className="mt-3 fs-4">
-                                                    Your submission is received and will be processed by the Marketplace! A CETAF administrator will review
-                                                    and score the taxonomic service. When the score is sufficient, the service will be published in the
-                                                    Marketplace catalog.
-                                                    <br />
-                                                    Thank you for using the Taxonomic Marketplace!
+                                                    {sended}
                                                 </p>
                                             }
                                         </Col>
@@ -71,7 +75,7 @@ const TaxonomicServiceForm = () => {
                                     {!completed &&
                                         <Row>
                                             <Col>
-                                                <FormBuilder formTemplate={TaxonomicServiceFormJSON}
+                                                <FormBuilder formTemplate={formTemplate}
                                                     SetCompleted={() => setCompleted(true)}
                                                 />
                                             </Col>
@@ -90,4 +94,4 @@ const TaxonomicServiceForm = () => {
     );
 };
 
-export default TaxonomicServiceForm;
+export default TaxonomicForm;
